ีupdated 147/150

+++ day3/v_uniq_package/main.cpp
@@ -8,13 +8,13 @@
 using namespace std;
 
 int main() {
-  CP::vector<float> v1;
+  CP::vector<int> v1;
   //read vector
   int n1,a;
   scanf("%d",&n1);
   v1.resize(n1);
   for (int i = 0;i < n1;i++) {
-    scanf("%f",&a);
+    scanf("%d",&a);
     v1[i] = a;
   }
 
@@ -23,7 +23,7 @@ int main() {
   //display result;
   printf("Result\n");
   for (auto &x : v1) {
-    printf("% ",x);
+    printf("%d ",x);
   }
   printf("\n");
